requirements/dev.txt
dj-database-url
django-cors-headers
django-dotenv
pillow
psycopg2